KP-8502 Created the layout structure for the app
Created elements for header, footer, and panels, that will be populated using portals.
Created a Panel component that renders a slide in panel on top of the content.
Created header and footer portal wrappers, and sample components.
Added a README file to the atoms directory describing the slots pattern.
Added sass style nesting support to tailwind styling.