KP-8888 Refactored how widgets manage their APIs (#48)
Updated the widget code to allow for using a mutable ref for defining an API object so that it can be mutated as the component state changes, which allows for the end user to always have access to the latest state.
Updated widgets to return a promise that is resolved with their API after the widget is initialized.