KP-9019 Added settings page (#58)
* KP-9019 Added settings page

Added a settings page with links to the profile, datastore, and any form of type Settings from the admin-center kapp.
Added a datastore UI that shows forms of type Data from the admin-center kapp and allows for seeing their submissions and rendering the forms.
Updated header to have a dropdown with links to the profile and settings page.

* KP-9019 Reverted source kapp for news and shortcuts

Reverted the news and shortcuts to continue fetching data from the catalog kapp to not break existing functionality.